+++ app/(routes)/product/[id]/_components/ProductContents.tsx
@@ -1,5 +1,5 @@
-import type { GetProductListResponse } from '@/app/modules/types';
-import { gruppo } from '@/app/modules/constants';
+import type { GetProductListResponse } from '@/app/_modules/types';
+import { gruppo } from '@/app/_modules/constants';
 import { Box } from '@mui/material';
 import clsx from 'clsx';
